Fix compile error

+++ src/textures/mlx_xpm42_utils.c
@@ -6,7 +6,7 @@
 /*   By: lde-la-h <lde-la-h@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/02/09 14:00:50 by lde-la-h      #+#    #+#                 */
-/*   Updated: 2022/02/11 11:56:14 by lde-la-h      ########   odam.nl         */
+/*   Updated: 2022/02/13 21:41:05 by W2Wizard      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ bool	mlx_insert_xpm_entry(t_xpm *xpm, char *line, uint32_t *ctable, size_t s)
 {
 	int32_t			index;
 
-	if (((uintptr_t)strrchr(line, ' ') - (uintptr_t)line) != xpm->cpp)
+	if (((uintptr_t)strrchr(line, ' ') - (uintptr_t)line) != (uint64_t)xpm->cpp)
 		return (false);
 	if (!isspace(line[xpm->cpp]) || line[xpm->cpp + 1] != '#' || \
 		!isalnum(line[xpm->cpp + 2]))
